requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
h11==0.13.0
httptools==0.4.0
idna==3.3
imageio==2.19.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
networkx==2.8.5
numpy==1.23.1
orjson==3.7.7
packaging==21.3
Pillow==9.2.0
pydantic==1.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
python-multipart==0.0.5
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tifffile==2022.5.4
transliterate==1.10.2
typing-extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
